+++ apps/common-app/src/examples/RuntimeTests/ReJest/stringFormatUtils.ts
