+++ apps/common-app/src/examples/RuntimeTests/RuntimeTestsExample.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
-import RuntimeTestsRunner from './ReanimatedRuntimeTestsRunner/RuntimeTestsRunner';
-import { describe } from './ReanimatedRuntimeTestsRunner/RuntimeTestsApi';
+import RuntimeTestsRunner from './ReJest/RuntimeTestsRunner';
+import { describe } from './ReJest/RuntimeTestsApi';
 
 export default function RuntimeTestsExample() {
   return (
