+++ apps/common-app/src/examples/RuntimeTests/tests/advancedAPI/measure.test.tsx
@@ -12,16 +12,8 @@ import Animated, {
   Easing,
   useFrameCallback,
 } from 'react-native-reanimated';
-import {
-  describe,
-  expect,
-  test,
-  render,
-  wait,
-  registerValue,
-  getRegisteredValue,
-} from '../../ReanimatedRuntimeTestsRunner/RuntimeTestsApi';
-import { ComparisonMode } from '../../ReanimatedRuntimeTestsRunner/types';
+import { describe, expect, test, render, wait, registerValue, getRegisteredValue } from '../../ReJest/RuntimeTestsApi';
+import { ComparisonMode } from '../../ReJest/types';
 
 describe('Test measuring component before nad after animation', () => {
   const INITIAL_MEASURE = 'INITIAL_MEASURE';
