+++ apps/common-app/src/examples/RuntimeTests/tests/advancedAPI/useFrameCallback.test.tsx
@@ -1,15 +1,7 @@
 import React, { useEffect } from 'react';
 import { View, StyleSheet } from 'react-native';
 import Animated, { useSharedValue, useAnimatedStyle, useFrameCallback } from 'react-native-reanimated';
-import {
-  describe,
-  test,
-  expect,
-  render,
-  getTestComponent,
-  wait,
-  useTestRef,
-} from '../../ReanimatedRuntimeTestsRunner/RuntimeTestsApi';
+import { describe, test, expect, render, getTestComponent, wait, useTestRef } from '../../ReJest/RuntimeTestsApi';
 
 describe('Test *****useFrameCallback*****', () => {
   describe('Test _canceling frameCallback_ after predefined time', () => {
