+++ apps/common-app/src/examples/RuntimeTests/tests/animations/withSpring/variousConfig.test.tsx
@@ -10,7 +10,7 @@ import {
   recordAnimationUpdates,
   render,
   wait,
-} from '../../../ReanimatedRuntimeTestsRunner/RuntimeTestsApi';
+} from '../../../ReJest/RuntimeTestsApi';
 import { Snapshots } from './withSpring.snapshot';
 
 const AnimatedComponent = ({
