+++ apps/common-app/src/examples/RuntimeTests/tests/layoutAnimations/entering/predefinedEntering.test.tsx
@@ -53,7 +53,7 @@ import {
   mockWindowDimensions,
   unmockWindowDimensions,
   waitForAnimationUpdates,
-} from '../../../ReanimatedRuntimeTestsRunner/RuntimeTestsApi';
+} from '../../../ReJest/RuntimeTestsApi';
 import {
   DurationEnteringSnapshots,
   NoModifierEnteringSnapshots,
