+++ apps/common-app/src/examples/RuntimeTests/tests/layoutAnimations/layout/predefinedLayoutPosition.test.tsx
@@ -20,7 +20,7 @@ import {
   getTestComponent,
   mockWindowDimensions,
   unmockWindowDimensions,
-} from '../../../ReanimatedRuntimeTestsRunner/RuntimeTestsApi';
+} from '../../../ReJest/RuntimeTestsApi';
 import {
   CurvedSnapshot,
   FadingSnapshot,
